Moved risklib.riskinput in commonlib
