+++ openquake/calculators/base.py
@@ -34,8 +34,8 @@ from openquake.hazardlib.geo import geodetic
 from openquake.baselib import general, hdf5
 from openquake.baselib.performance import Monitor
 from openquake.hazardlib.calc.filters import RtreeFilter
-from openquake.risklib import riskinput, __version__ as engine_version
-from openquake.commonlib import readinput, riskmodels, datastore, source
+from openquake.risklib import __version__ as engine_version
+from openquake.commonlib import readinput, datastore, source, riskinput
 from openquake.commonlib.oqvalidation import OqParam
 from openquake.commonlib.parallel import starmap, executor, wakeup_pool
 from openquake.baselib.python3compat import with_metaclass
