+++ openquake/calculators/event_based_risk.py
@@ -28,8 +28,8 @@ from openquake.baselib import hdf5
 from openquake.baselib.python3compat import zip
 from openquake.baselib.general import AccumDict, humansize, block_splitter
 from openquake.calculators import base, event_based
-from openquake.commonlib import parallel, calc
-from openquake.risklib import riskinput, scientific
+from openquake.commonlib import parallel, calc, riskinput
+from openquake.risklib import scientific
 from openquake.commonlib.parallel import starmap
 
 U32 = numpy.uint32
