+++ openquake/commonlib/export/hazard.py
@@ -33,7 +33,7 @@ from openquake.hazardlib.calc import disagg, gmf
 from openquake.commonlib.export import export
 from openquake.commonlib.writers import write_csv
 from openquake.commonlib import writers, hazard_writers, util
-from openquake.risklib.riskinput import GmfGetter
+from openquake.commonlib.riskinput import GmfGetter
 from openquake.commonlib import calc
 
 F32 = numpy.float32
