+++ openquake/commonlib/readinput.py
@@ -33,11 +33,11 @@ from openquake.baselib.python3compat import configparser, encode
 from openquake.baselib import hdf5
 from openquake.hazardlib import geo, site, imt
 from openquake.hazardlib.calc.hazard_curve import zero_curves
-from openquake.risklib import riskmodels, riskinput, valid
+from openquake.risklib import riskmodels, valid
 from openquake.commonlib import datastore
 from openquake.commonlib.oqvalidation import OqParam
 from openquake.commonlib.node import Node, context
-from openquake.commonlib import nrml, logictree, InvalidFile
+from openquake.commonlib import nrml, logictree, InvalidFile, riskinput
 from openquake.commonlib.riskmodels import get_risk_models
 from openquake.commonlib import source, sourceconverter
