+++ openquake/commonlib/riskinput.py
